fix extra digit in date input
