+++ ckanext/dalrrd_emc_dcpr/helpers.py
@@ -1,6 +1,7 @@
 import json
 import logging
 import typing
+import datetime
 from urllib.parse import quote
 from html import escape as html_escape
 
@@ -13,7 +14,6 @@ from . import constants
 from .logic.action.emc import show_version
 from .constants import DCPRRequestStatus
 from .model.dcpr_request import DCPRRequest
-
 
 logger = logging.getLogger(__name__)
 
@@ -345,7 +345,6 @@ def get_dcpr_requests_approved_by_nsif(request_origin):
 
 
 def is_dcpr_request(package):
-
     for extra in package.get("extras") or []:
         if extra.get("key") == "origin" and extra.get("value") == "DCPR":
             return True
@@ -353,8 +352,11 @@ def is_dcpr_request(package):
 
 
 def get_dcpr_request_action(package):
-
     for extra in package.get("extras") or []:
         if extra.get("key") == "action" and extra.get("value") == "APPROVE":
             return "ACCEPT"
     return "REJECT"
+
+
+def get_today_date() -> str:
+    return datetime.datetime.now().strftime("%Y-%m-%d")
